+++ src/translations/fr.ts
@@ -5,12 +5,6 @@ export const fr: Translations = {
     total: 'Total',
     grid: 'Réseau',
   },
-  grid_coupling: {
-    title_default: 'Couplage réseau',
-  },
-  pv_system: {
-    title_default: 'Production photovoltaïque',
-  },
   confirm: {
     toggle: 'Êtes-vous sûr de vouloir commuter « {name} » ?',
   },